refactor(android): merge supportui and rename folder
