+++ android/sdk/gradle.properties
@@ -121,7 +121,6 @@ EXCLUDE_LIBRARY_FILES=
 
 #features
 ENABLE_SO_DOWNLOAD=false
-INCLUDE_SUPPORT_UI=true
 
 #maven central publish params
 PUBLISH_ARTIFACT_ID=hippy-common
