+++ android/sdk/src/main/java/androidx/recyclerview/widget/ItemLayoutParams.java
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.RecyclerView.LayoutParams;
  * Created on 2021/3/17.
  * Description
  */
-public interface IItemLayoutParams {
+public interface ItemLayoutParams {
 
     void getItemLayoutParams(int position, LayoutParams lp);
 }
