+++ android/sdk/src/main/java/com/tencent/mtt/hippy/views/hippylist/HippyRecyclerViewWrapper.java
@@ -29,7 +29,7 @@ import com.tencent.mtt.hippy.HippyEngineContext;
 import com.tencent.mtt.hippy.HippyInstanceContext;
 import com.tencent.mtt.hippy.uimanager.HippyViewBase;
 import com.tencent.mtt.hippy.uimanager.NativeGestureDispatcher;
-import com.tencent.mtt.nxeasy.recyclerview.helper.skikcy.IHeaderHost;
+import com.tencent.mtt.hippy.views.hippylist.recyclerview.helper.skikcy.IHeaderHost;
 
 /**
  * Description
