+++ android/sdk/src/main/java/com/tencent/mtt/hippy/views/hippylist/PullHeaderEventHelper.java
@@ -24,10 +24,10 @@ import com.tencent.mtt.hippy.common.HippyMap;
 import com.tencent.mtt.hippy.uimanager.HippyViewEvent;
 import com.tencent.mtt.hippy.uimanager.PullHeaderRenderNode;
 import com.tencent.mtt.hippy.utils.PixelUtil;
-import com.tencent.mtt.nxeasy.recyclerview.helper.header.HeaderRefreshHelper;
-import com.tencent.mtt.nxeasy.recyclerview.helper.header.IHeaderRefreshListener;
-import com.tencent.mtt.nxeasy.recyclerview.helper.header.IHeaderRefreshView;
-import com.tencent.mtt.nxeasy.recyclerview.helper.header.ILayoutRequester;
+import com.tencent.mtt.hippy.views.hippylist.recyclerview.helper.header.HeaderRefreshHelper;
+import com.tencent.mtt.hippy.views.hippylist.recyclerview.helper.header.IHeaderRefreshListener;
+import com.tencent.mtt.hippy.views.hippylist.recyclerview.helper.header.IHeaderRefreshView;
+import com.tencent.mtt.hippy.views.hippylist.recyclerview.helper.header.ILayoutRequester;
 
 public class PullHeaderEventHelper implements IHeaderRefreshListener, IHeaderRefreshView,
         ILayoutRequester {
